feat: Record 액션 편집/프리셋, HandDetail 마킹, Players 모듈 분리, UI 개선
- Record: EditActionSheet, BoardCardsSection, ActionPresetCalculator 추가, StreetStep 편집 기능
- HandDetail: TopBar 엔드 콘텐츠 분리, MarkPlayerOnHandUseCase 연동, ResultSection 개선
- Players: players-edit 모듈 분리, PlayerEditSheet 이동
- Core: AmountVisualTransformation 추가, Badge 제거, ThemePreviews 정리
- iOS: StatusBarController 추가, ContentView/Info.plist 수정
- 다국어 문자열 리소스 추가, 테스트 보강

+++ feature/hand-detail/src/commonMain/kotlin/com/hand/log/handdetail/HandDetailContent.kt
@@ -39,7 +39,7 @@ internal fun HandDetailContent(
 	hand: HandRecord,
 	useBbUnit: Boolean,
 	graphicsLayer: GraphicsLayer,
-	onMarkPlayer: () -> Unit = {},
+	onMarkPlayer: (Int) -> Unit = {},
 	modifier: Modifier = Modifier,
 ) {
 	val colors = HandyTheme.colorScheme
